Fixed Mounting ISO images,loads HCRT.DBG.Z properly (#1)
* Fixed Mounting ISO images,loads HCRT.DBG.Z properly

The issue with mounting ISO images was that in TempleOS,the filenames must be
38 chars long and this was affecting `MountFile`.

It also now loads the debugger into after we load the imports(we attach the dbg info to existing symbols).

I will make a guide on building a TempleOS distro on tos.

Nrootconauto

* dbg.cxx works on FreeBSD,misc changes

Removed useless function and vfs.cxx updated to match TempleOS filename length

* Made HCRT.DBG.Z cannon,updated HCRT_BOOTSTRAP.BIN

Removed misleading vfs.cxx comment leftover from last commit.

Nrootconauto

* Files not needed for bootrapping,removed. Nrootconauto

* ???

+++ dbg.cxx
@@ -111,7 +111,7 @@ static void routine(int sig, siginfo_t* info, ucontext_t* ctx) {
       REG(EFL),
   };
 #elif defined(__FreeBSD__)
-#define REG(X) ctx->uc_mcontext.mc_##X
+#define REG(X) static_cast<uint64_t>(ctx->uc_mcontext.mc_##X)
   uint64_t regs[] = {
       REG(rax),    REG(rcx), REG(rdx),
       REG(rbx),    REG(rsp), REG(rbp),
